chore: move and rename splitter wrapper
