+++ src/skrutable/splitting.py
